backend/customer-ms/src/main/java/com/fbs/customer/model/User.java
package com.fbs.customer.model;	


import org.springframework.data.mongodb.core.mapping.Document;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;



@Document(collection = "users")
public class User {
	
	@NotBlank
	@Email
	private String username;
	
	@NotBlank
	@Size(min = 8, max=20)
	private String password;
	@NotBlank
	private String role;

	public User(String username, String password, String role) {
		super();
		this.username = username;
		this.password = password;
		this.role = role;
	}

	public User() {
		super();
	}

	public String getUsername() {
		return username;
	}

	public void setUsername(String username) {
		this.username = username;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}

	public String getRole() {
		return role;
	}

	public void setRole(String role) {
		this.role = role;
	}

}
